requirements.txt
click==8.0.3
deepchem==2.5.0
Flask==2.1.1
gunicorn==20.1.0
numpy==1.22.3
pandas==1.4.2
rdkit-pypi==2021.9.2.1
scikit-learn==0.24.2
tensorflow-cpu==2.6.1
Werkzeug==2.0.2
